# RailWayStation/docs/swagger.yaml
basePath: /api
definitions:
  models.Area:
    properties:
      districts:
        items:
          $ref: '#/definitions/models.District'
        type: array
      id:
        type: integer
      name:
        type: string
      region:
        $ref: '#/definitions/models.Region'
      region_id:
        type: integer
    type: object
...
  models.District:
    properties:
      area:
        $ref: '#/definitions/models.Area'
      area_id:
        type: integer
      id:
        type: integer
      name:
        type: string
      rail_stations:
        items:
          $ref: '#/definitions/models.RailStation'
        type: array
    type: object
...
  models.Line:
    properties:
      id:
        type: string
      name:
        type: string
      rail_stations:
        items:
          $ref: '#/definitions/models.RailStation'
        type: array
    type: object
...
  models.RailStation:
    properties:
      district:
        $ref: '#/definitions/models.District'
      district_id:
        type: integer
      hash:
        type: string
      id:
        type: integer
      lat:
        type: number
      line:
        $ref: '#/definitions/models.Line'
      line_id:
        type: string
      lon:
        type: number
      name:
        type: string
    type: object
...
  models.Region:
    properties:
      areas:
        items:
          $ref: '#/definitions/models.Area'
        type: array
      id:
        type: integer
      name:
        type: string
    type: object
  models.RegionCreateRequest:
    properties:
      name:
        maxLength: 100
        minLength: 2
        type: string
    required:
    - name
    type: object
  models.RegionUpdateRequest:
    properties:
      name:
        maxLength: 100
        minLength: 2
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API для управления железнодорожными станциями
  title: Railway Stations API
  version: "1.0"
paths:
...
  /regions:
    get:
      description: Get a list of all regions with pagination and filtering
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Number of items per page
        in: query
        name: limit
        type: integer
      - description: Filter by name
        in: query
        name: name
        type: string
      - default: id
        description: Sort field
        in: query
        name: sort_by
        type: string
      - default: asc
        description: Sort order
        enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get all regions
      tags:
      - regions
    post:
      consumes:
      - application/json
      description: Create a new region in the database
      parameters:
      - description: Region object
        in: body
        name: region
        required: true
        schema:
          $ref: '#/definitions/models.RegionCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Region'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new region
      tags:
      - regions
  /regions/{id}:
    delete:
      description: Delete a region by ID
      parameters:
      - description: Region ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Delete region
      tags:
      - regions
    get:
      description: Get a single region by its ID
      parameters:
      - description: Region ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Region'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get region by ID
      tags:
      - regions
    put:
      consumes:
      - application/json
      description: Update an existing region by ID
      parameters:
      - description: Region ID
        in: path
        name: id
        required: true
        type: integer
      - description: Region update object
        in: body
        name: region
        required: true
        schema:
          $ref: '#/definitions/models.RegionUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Region'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Update region
      tags:
      - regions
...
swagger: "2.0"